feat: Allow adding nodes via canvas click
Introduces a new feature that enables users to add nodes by clicking on the Blueprint canvas. This improves the user experience by providing a more direct and intuitive way to interact with the visual graph. The `BlueprintVisualizer` now accepts an `onCanvasClick` prop, which is triggered when a click occurs on the canvas background. This event data is then used by the `BlueprintCreator` component to present a node picker at the clicked location.